fix(onboard): flag unsupported DOCKER_HOST endpoints even when reachable
Address review on the invalid-DOCKER_HOST advisory:

- A reachable tcp:// or ssh:// DOCKER_HOST still cannot back the managed
  gateway (the gateway env writer only serializes an absolute unix://
  socket), so onboarding must block it rather than proceed and fail
  later. Drop the reachability guard on invalid_docker_host and reword
  the advisory to state the endpoint is unsupported even when reachable.
- Move the pure isSupportedGatewayDockerHost predicate into the domain
  layer (src/lib/domain/docker-host.ts) per the domain-purity contract,
  imported by both preflight and the gateway env writer.
- Reject null bytes and line breaks in the raw DOCKER_HOST value before
  trimming, so a trailing newline can no longer be trimmed away and then
  accepted; the socket path still rejects the single quote it is wrapped
  in when written to the gateway environment file.

Refs #7731
