+++ src/lib/onboard/docker-driver-gateway-env.test.ts
@@ -10,34 +10,9 @@ import { describe, expect, it, vi } from "vitest";
 import {
   buildDockerDriverGatewayEnv,
   buildDockerGatewayDebEnvFile,
-  isSupportedGatewayDockerHost,
   startPackageManagedDockerDriverGatewayWithEnvOverride,
   writeDockerGatewayDebEnvOverride,
 } from "./docker-driver-gateway-env";
-
-describe("isSupportedGatewayDockerHost (#7731)", () => {
-  it.each([
-    ["unset", undefined],
-    ["empty", ""],
-    ["whitespace only", "   "],
-    ["absolute unix socket", "unix:///var/run/docker.sock"],
-    ["absolute unix socket with surrounding whitespace", "  unix:///var/run/docker.sock  "],
-  ])("accepts %s DOCKER_HOST", (_case, value) => {
-    expect(isSupportedGatewayDockerHost(value)).toBe(true);
-  });
-
-  it.each([
-    ["a TCP endpoint", "tcp://203.0.113.10:2375"],
-    ["an ssh endpoint", "ssh://user@host"],
-    ["an fd endpoint", "fd://"],
-    ["a bare socket path without a scheme", "/var/run/docker.sock"],
-    ["a relative unix path", "unix://relative/docker.sock"],
-    ["a unix socket path with a quote", "unix:///var/run/dock'er.sock"],
-    ["a unix socket path with an embedded newline", "unix:///var/run/\ndocker.sock"],
-  ])("rejects %s", (_case, value) => {
-    expect(isSupportedGatewayDockerHost(value)).toBe(false);
-  });
-});
 
 function homeEnv(home: string, xdgConfigHome = ""): NodeJS.ProcessEnv {
   return { HOME: home, XDG_CONFIG_HOME: xdgConfigHome } as NodeJS.ProcessEnv;
