+++ src/lib/onboard/docker-driver-gateway-env.ts
@@ -13,6 +13,7 @@ import {
   WILDCARD_GATEWAY_BIND_ADDRESS,
 } from "../core/gateway-address";
 import { DEFAULT_GATEWAY_PORT, GATEWAY_PORT } from "../core/ports";
+import { isSupportedGatewayDockerHost } from "../domain/docker-host";
 import {
   DOCKER_DRIVER_GATEWAY_JWT_TTL_SECS,
   prepareDockerDriverGatewayConfigEnv,
@@ -251,20 +252,6 @@ function formatEnvironmentFileAssignment(key: string, value: string): string {
     return `${key}='${dockerHost}'`;
   }
   return `${key}=${value}`;
-}
-
-// A DOCKER_HOST value onboarding can use. Unset means Docker's default socket,
-// which is supported; a set value must be an absolute `unix://` socket with no
-// serialization-hostile characters. TCP endpoints and relative paths are
-// unsupported, so when DOCKER_HOST points at one, `docker info` cannot reach the
-// daemon and onboarding treats the host as invalid.
-export function isSupportedGatewayDockerHost(value: string | undefined): boolean {
-  const candidate = String(value ?? "").trim();
-  if (!candidate) return true;
-  const prefix = "unix://";
-  if (!candidate.startsWith(prefix)) return false;
-  const socketPath = candidate.slice(prefix.length);
-  return path.isAbsolute(socketPath) && !/[\0\r\n']/.test(socketPath);
 }
 
 function normalizePackageServiceDockerHost(value: string | undefined): string | undefined {
